Improve member creation and store member image

+++ app/Http/Controllers/Admin/MemberController.php
@@ -8,6 +8,15 @@ use App\Models\Member;
 
 class MemberController extends Controller
 {
+    protected $limit = 10;
+
+    protected $uploadPath;
+
+    public function __construct()
+    {
+       $this->uploadPath = public_path('members');
+    }
+
     public function index(){
         $members = Member::latest()->paginate(10);
         return view('admin.members.index',compact('members'));
@@ -19,9 +28,10 @@ class MemberController extends Controller
     }
 
     public function store(MemberRequest $request){
-        $member = Member::create($request->all());
+        $data = $this->handleRequest($request);
 
-        $this->storeImage($member);
+        Member::create($data);
+
         return redirect()->route('admin.members.index')->with('success','New member added successfully!');
     }
 
@@ -30,10 +40,12 @@ class MemberController extends Controller
     }
 
     public function update(MemberRequest $request,Member $member){
-        $oldImage = $member->image;
-        $member->update($request->all());
 
-        $this->storeImage($member);
+        $data = $this->handleRequest($request);
+
+        $oldImage = $member->image;
+
+        $member->update($data);
 
         if($oldImage !== $member->image){
             $this->removeImage($oldImage);
@@ -48,21 +60,28 @@ class MemberController extends Controller
         return redirect()->route('admin.members.index')->with('success','Member deleted successfully!');
     }
 
-    private function storeImage($member)
-    {
-        if (request()->has('image')) {
-            $member->update([
-                'image' => request()->image->store('members', 'public'),
-            ]);
+
+    private function handleRequest($request){
+        $data = $request->all();
+
+        if($request->hasFile('image')){
+            $image = $request->file('image');
+            $fileName = $image->getClientOriginalName();
+            $destination = $this->uploadPath;
+
+            $image->move($destination,$fileName);
+            
+            $data['image'] = $fileName;
         }
+
+        return $data;
     }
 
     private function removeImage($image){
         if(!empty($image)){
-            $imagePath = 'storage/'.$image;
-            if(file_exists($imagePath)){
-                unlink($imagePath);
-            }
+            $imagePath = $this->uploadPath.'/'.$image;
+
+            if(file_exists($imagePath)) unlink($imagePath);
         }
     }
 }
